Rename progress tracker files
