+++ test/e2e/conftest.py
@@ -14,7 +14,7 @@ from layer.config import ClientConfig, Config, ConfigManager
 from layer.contracts.fabrics import Fabric
 from layer.contracts.projects import Project
 from layer.projects.project_runner import ProjectRunner
-from layer.tracker.remote_execution_project_progress_tracker import (
+from layer.tracker.remote_execution_progress_tracker import (
     RemoteExecutionRunProgressTracker,
 )
 from test.e2e.assertion_utils import E2ETestAsserter
@@ -142,7 +142,7 @@ def _cleanup_project(client: LayerClient, project: Project):
 @pytest.fixture()
 def project_runner(config: Config) -> ProjectRunner:
     return ProjectRunner(
-        config, project_progress_tracker_factory=RemoteExecutionRunProgressTracker
+        config, progress_tracker_factory=RemoteExecutionRunProgressTracker
     )
